refactored model classes
