+++ models/__init__.py
@@ -1,9 +1,2 @@
 # -*- coding: utf-8 -*-
-
-
-from models.lstm import LSTM
-from models.senticgcn import SenticGCN
-from models.attsenticgcn import ATTSenticGCN
-from models.sdgcn import SDGCN
-from models.affectivegcn import AFFGCN
-from models.senticgcn_bert import SenticGCN_BERT
+from models.model import AFFGCN, ATTSenticGCN, LSTM, SDGCN, SenticGCN, BERT_SPC, SenticGCN_BERT, SenticGCNGLOVE
