--- models/senticgcn_bert.py
@@ -1,104 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import math
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from layers.dynamic_rnn import DynamicLSTM
-from .posf import position_weight
-
-class GraphConvolution(nn.Module):
-    """
-    Simple GCN layer, similar to https://arxiv.org/abs/1609.02907
-    """
-    def __init__(self, in_features, out_features, bias=True):
-        super(GraphConvolution, self).__init__()
-        self.in_features = in_features
-        self.out_features = out_features
-        self.weight = nn.Parameter(torch.FloatTensor(in_features, out_features))
-        if bias:
-            self.bias = nn.Parameter(torch.FloatTensor(out_features))
-        else:
-            self.register_parameter('bias', None)
-
-    def forward(self, text, adj):
-        text = text.to(torch.float32)
-        hidden = torch.matmul(text, self.weight)
-        denom = torch.sum(adj, dim=2, keepdim=True) + 1
-        output = torch.matmul(adj, hidden) / denom
-        if self.bias is not None:
-            return output + self.bias
-        else:
-            return output
-
-class SenticGCN_BERT(nn.Module):
-    def __init__(self, bert, opt):
-        super(SenticGCN_BERT, self).__init__()
-        self.opt = opt
-        self.bert = bert
-        #self.dropout = nn.Dropout(opt.dropout)
-        #self.embed = nn.Embedding.from_pretrained(torch.tensor(embedding_matrix, dtype=torch.float))
-        #self.text_lstm = DynamicLSTM(768, opt.hidden_dim, num_layers=1, batch_first=True, bidirectional=True)
-        self.gc1 = GraphConvolution(opt.bert_dim, opt.bert_dim)
-        self.gc2 = GraphConvolution(opt.bert_dim, opt.bert_dim)
-        self.gc3 = GraphConvolution(opt.bert_dim, opt.bert_dim)
-        #self.gc4 = GraphConvolution(2*opt.hidden_dim, 2*opt.hidden_dim)
-        #self.gc5 = GraphConvolution(2*opt.hidden_dim, 2*opt.hidden_dim)
-        #self.gc6 = GraphConvolution(2*opt.hidden_dim, 2*opt.hidden_dim)
-        #self.gc7 = GraphConvolution(2*opt.hidden_dim, 2*opt.hidden_dim)
-        #self.gc8 = GraphConvolution(2*opt.hidden_dim, 2*opt.hidden_dim)
-        self.fc = nn.Linear(opt.bert_dim, opt.polarities_dim)
-        self.text_embed_dropout = nn.Dropout(0.3)
-    
-    def position_weight(self, x, aspect_double_idx, text_len, aspect_len):
-        return position_weight(self.opt, x, aspect_double_idx, text_len, aspect_len)
-
-    def mask(self, x, aspect_double_idx):
-        batch_size, seq_len = x.shape[0], x.shape[1]
-        aspect_double_idx = aspect_double_idx.cpu().numpy()
-        mask = [[] for i in range(batch_size)]
-        for i in range(batch_size):
-            for j in range(aspect_double_idx[i,0]):
-                mask[i].append(0)
-            for j in range(aspect_double_idx[i,0], min(aspect_double_idx[i,1]+1, self.opt.max_seq_len)):
-                mask[i].append(1)
-            for j in range(min(aspect_double_idx[i,1]+1, self.opt.max_seq_len), seq_len):
-                mask[i].append(0)
-        mask = torch.tensor(mask).unsqueeze(2).float().to(self.opt.device)
-        return mask*x
-
-    def forward(self, inputs):
-        text_bert_indices, text_indices, aspect_indices, bert_segments_ids, left_indices, adj = inputs
-        text_len = torch.sum(text_indices != 0, dim=-1)
-        aspect_len = torch.sum(aspect_indices != 0, dim=-1)
-        left_len = torch.sum(left_indices != 0, dim=-1)
-        aspect_double_idx = torch.cat([left_len.unsqueeze(1), (left_len+aspect_len-1).unsqueeze(1)], dim=1)
-        #text = self.embed(text_indices)
-        #text = self.text_embed_dropout(text)
-        #text_out, (_, _) = self.text_lstm(text, text_len)
-
-        #encoder_layer, pooled_output = self.bert(text_bert_indices, token_type_ids=bert_segments_ids, output_all_encoded_layers=False)
-        od = self.bert(text_bert_indices, token_type_ids=bert_segments_ids, output_hidden_states=None)
-        encoder_layer = od['last_hidden_state']
-        pooled_output = od['pooler_output']
-
-        text_out = encoder_layer
-
-        x = F.relu(self.gc1(self.position_weight(text_out, aspect_double_idx, text_len, aspect_len), adj))
-        x = F.relu(self.gc2(self.position_weight(x, aspect_double_idx, text_len, aspect_len), adj))
-        #x = F.relu(self.gc1(text_out, adj))
-        #x = F.relu(self.gc2(x, adj))
-        x = F.relu(self.gc3(self.position_weight(x, aspect_double_idx, text_len, aspect_len), adj))
-        #x = F.relu(self.gc4(self.position_weight(x, aspect_double_idx, text_len, aspect_len), adj))
-        #x = F.relu(self.gc5(self.position_weight(x, aspect_double_idx, text_len, aspect_len), adj))
-        #x = F.relu(self.gc6(self.position_weight(x, aspect_double_idx, text_len, aspect_len), adj))
-        #x = F.relu(self.gc7(self.position_weight(x, aspect_double_idx, text_len, aspect_len), adj))
-        #x = F.relu(self.gc8(self.position_weight(x, aspect_double_idx, text_len, aspect_len), adj))
-
-        # commented out by CP John @ Nov 2
-        # x = self.mask(x, aspect_double_idx)
-        alpha_mat = torch.matmul(x, text_out.transpose(1, 2))
-        alpha = F.softmax(alpha_mat.sum(1, keepdim=True), dim=2)
-        x = torch.matmul(alpha, text_out).squeeze(1) # batch_size x 2*hidden_dim
-        output = self.fc(x)
-        return output
